refactor(cli): address review feedback on the workload split
- Move runHooks and errHookFailed out of wait.go into a dedicated hooks.go, and move errStdinFDOverflow to debug.go where it is used; wait.go is now just the wait command.
- Use the shared permissionWrite constant in rollout.go and create.go (they set the annotation via map index, which the PR2 sed missed).
- Correct the pollForChanges comment: a blocking send guarantees at least one apply, not one-for-one delivery, since later events can coalesce.
